Create projectGradient.html
This project is a simple web application that dynamically generates a gradient background using randomly chosen colors. The application consists of two main buttons, each responsible for generating a random color. When either button is clicked, the background gradient is updated, and the CSS code for the gradient is displayed.